Extracted getting build logic script and turned optional
